Introduce `RuntimeStrategy`
Determine the strategy to take when selecting a runtime to work with a cluster.
The problem is that we want to select a runtime to match an existing cluster AND
we want to have a choice in place for what to do when there is no cluster yet.

This work is not yet complete. For example, `Cluster::new` now determines the
runtime to use immediately. However, this conflicts with the functions in
`coordinate`, i.e. it would be more appropriate to make that determination
_after_ entering the locked regions of `coordinate`'s functions.